requirements.txt
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.3
arrow==0.10.0
asn1crypto==0.22.0
billiard==3.5.0.2
blessed==1.14.2
celery==4.0.2
certifi==2017.1.23
cffi==1.10.0
click==6.7
cryptography==1.8.1
Django==1.9.12
django-celery-results==1.0.1
django-mptt==0.8.7
django-picklefield==0.3.2
enum34==1.1.6
future==0.16.0
gevent==1.2.1
greenlet==0.4.12
idna==2.5
ipaddress==1.0.18
kombu==4.0.2
lxml==3.7.3
packaging==16.8
psycopg2==2.7.1
pycparser==2.17
pyOpenSSL==16.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
six==1.10.0
urllib3==1.20
vine==1.1.3
wcwidth==0.1.7
